+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 * Fixed an issue where map loading would not include some additional models in some cases.
 * Fixed an issue where xc3_tex would not create the output folder for the `edit-wimdo`, `edit-wilay`, and `edit-wifnt` commands.
+* Fixed an issue where xc3_model would not use the correct material alpha test texture in some cases.
 
 ## 0.21.0 - 2025-11-25
 ### Added